+++ fivox/frequencyFunctor.h
@@ -37,11 +37,16 @@ class FrequencyFunctor : public EventFunctor< TImage >
     typedef typename Super::TSpacing TSpacing;
 
 public:
-    FrequencyFunctor() {}
+    FrequencyFunctor( const float magnitude )
+        : _magnitude( magnitude )
+    {}
     virtual ~FrequencyFunctor() {}
 
     TPixel operator()( const TPoint& point, const TSpacing& spacing )
         const override;
+
+private:
+    const float _magnitude;
 };
 
 template< class TImage > inline typename FrequencyFunctor< TImage >::TPixel
@@ -65,7 +70,7 @@ FrequencyFunctor< TImage >::operator()( const TPoint& itkPoint,
 
     float sum = 0.f;
     for( const Event& event : events )
-        sum = std::max( sum, event.value );
+        sum = std::max( sum, event.value * _magnitude );
     return Super::_scale( sum );
 }